✅ Add tests
